feat: Add Rish-based installer with signature verification and downgrade support
Hello, This is my first PR and I'm so excited about it.

# [What is rish](https://github.com/RikkaApps/Shizuku-API/blob/master/rish/README.md)
rish is an Android program for interacting with a shell that runs on a high-privileged daemon process. In a way, it allows to execute adb shell commands

A typical `rish` installation allows the user to use `rish` commands from everywhere by putting it inside `usr/bin/`

# What I did
I added `rish` support to Revancify in a way that allows installation of apps directly from the terminal, uninstall apps, get installed versions, in a similar way that the current `su` functionalities do.

# Details
- I added another flag `-R` to disable `rish`
- Inside `install.sh`, I modified so that when we have `rish` access, we'll use it.
- Inside `root.sh`, I added `rish` functionality to `getInstalledVersion`, so It'll work even without `su`, useful to check the current installation in the selection screen.
- Added a new `modules/system/rish.sh` that will handle the installation.
- Now `LAUNCH_APP_AFTER_MOUNT` works without su, it's the same implementation, but using `am force-stop` to stop the Keyboard and Termux instead of `kill -9`

# Installation Details
I verify the signature of the current installation (if there is), using `keytool`, so if there is a signature mismatch, the user can decide to continue (uninstalling the current app) or abort:
```
local STOCK_APP_PATH=$(rish -c 'pm path "'"$PKG_NAME"'" | sed -n "/base/s/package://p"')
local STOCK_APP_SIGNATURE=$(keytool -printcert -jarfile "$STOCK_APP_PATH" 2>/dev/null | awk '/SHA256:/{print $2}' | tr -d ':')
```

I check if the installation is a downgrade using the same command used in `import.sh`
```
    if jq -e '.[0] > .[1]' <<< "[\"${INSTALLED_VERSION:-0}\", \"$APP_VER\"]" &> /dev/null; then
```
The program checks if the user has set `ALLOW_APP_VERSION_DOWNGRADE` in Revancify configs, and decides to proceed or abort.

Choosing to continue with the downgrade, the program will try to install with two ways, using the `-d` flag from the [pm install](https://developer.android.com/tools/adb#pm) command (although it has never worked for me). If it fails, the script will try uninstalling the app and installing again.

If it isn't a downgrade or the stock app is installed, it'll install normally.

> Similar to the `su` version, the program uses mount_log.txt to log details.

# Testing
✅ Installed new app via rish
✅ Updated existing app with matching signature
✅ Prompted uninstall on signature mismatch
✅ Handled downgrade with config on/off
✅ Verified log creation and deletion of temp APK

# Video
https://github.com/user-attachments/assets/53bcb6ca-9bd7-4425-b8b7-665dda6ea60f