feat: Add Rish-based installer with signature verification and downgrade support  (#2)
# New changes
I reorganized the code to be even more robust, with a similar structure than `su`:
- User queries and logic is in `.modules/system/rish.sh`
- Renamed `.root/` to `.system/` and changed the paths of the new scripts inside `.modules/system/root.sh` accordingly
- The install and uninstall scripts are in `.system/rish-install.sh` and `.system/rish-uninstall.sh`, they are only there to execute the functions and verify.
- Added more queries to the user.

# Tested workflows
✅ Installed new app
✅ Updated existing app with matching signature
✅ Prompted uninstall on signature mismatch, ask permission to te user to uninstall current installation
- ✅ User says yes
- ✅ User says no, abort
✅ Handled downgrade with config on/off, if downgrades are allowed,  ask permission to te user to uninstall current installation
- ✅ User says yes
- ✅ User says no, abort
✅ Verified log creation and deletion of temp APK

# Multiple users installations
When I was testing with some apps, I found some weird errors related to more users, it was because I had the same package name installed in my secure folder. So if one of the workflows above fails (downgrade or signature mismatch) , I checked for hidden apps with `dumpsys package <package>`, and prompted the user for permission to uninstall the app from all users.
![Captura desde 2025-05-30 08-17-26](https://github.com/user-attachments/assets/a6c5190a-f565-4833-9971-7f567ac2c0c1)

### Why not uninstall from all users directly?
Some apps installed directly from my secured folder couldn't be uninstalled with `rish` , and required manual uninstallation from the user. Also some times I could have a stock app in my secure folder and a patched app in my normal user, but most of the time I couldn't have mismatched apps in each user.
So I found it very inconsistent, so I did that if we encounter a hidden app, we do a second attempt if the user allows it, if we can leave the other users intact, the better.

# Video
In this video I have in both my user and in the secure folder the stock PlayStore version or RAR:

https://github.com/user-attachments/assets/5ed6c3eb-f517-4d8f-9258-6c724cb2c6ea